projects/trading-system/code/closed-system-trading/requirements.txt
python-binance
python-dotenv
supabase
streamlit
pandas
ta
